+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Management of physical channel: the actual closing is now done by the card removal procedure. The associated timeout
+  has been removed.
+### Removed
+- Dependency to logger implementation.
 
 ## [2.1.0] - 2023-11-13
 :warning: **CAUTION**: this version requires to use at least version `2.3.2` of the